Added logic for creating tables and storing embeddings
