+++ tables/create_table_and_vectorize/postgres_setup_trail.py
