+++ tables/create_table_and_vectorize/table_creation_and_storing.txt
