+++ tables/create_table_and_vectorize/vectorize_sample_repo_simple.py
