+++ tables/create_table_and_vectorize/view_vector.py
